Split validators module
