+++ ldap2pg/config.py
@@ -20,8 +20,6 @@ import sys
 import yaml
 
 from . import __version__
-from .acl import Acl
-from .ldap import parse_scope
 from .utils import (
     deepget,
     deepset,
@@ -29,7 +27,7 @@ from .utils import (
     string_types,
     make_group_map,
 )
-from .role import RoleOptions
+from . import validators as V
 
 
 logger = logging.getLogger(__name__)
@@ -69,212 +67,6 @@ class ColoredStreamHandler(logging.StreamHandler):
             for line in lines.splitlines(True)
         ])
         return lines
-
-
-def acldict(value):
-    if not hasattr(value, 'items'):
-        raise ValueError('acl_dict must be a dict')
-
-    return dict([
-        (k, Acl(k, **v))
-        for k, v in value.items()
-    ])
-
-
-def raw(v):
-    return v
-
-
-def ldapquery(value):
-    query = dict(Configuration.DEFAULTS['ldap']['default_query'], **value)
-
-    if 'attribute' in query:
-        query['attributes'] = query['attribute']
-        del query['attribute']
-    if isinstance(query['attributes'], string_types):
-        query['attributes'] = [query['attributes']]
-
-    query['scope'] = parse_scope(query['scope'])
-
-    return query
-
-
-def rolerule(value):
-    rule = value
-
-    if value is None:
-        raise ValueError("Empty role rule. Wrong indentation?")
-
-    if isinstance(rule, string_types):
-        rule = dict(names=[rule])
-
-    if 'name' in rule:
-        rule['names'] = rule.pop('name')
-    if 'names' in rule and isinstance(rule['names'], string_types):
-        rule['names'] = [rule['names']]
-
-    if 'names' not in rule and 'name_attribute' not in rule:
-        raise ValueError("Missing role name")
-
-    if 'parent' in rule:
-        rule['parents'] = rule.pop('parent')
-    rule.setdefault('parents', [])
-    if isinstance(rule['parents'], string_types):
-        rule['parents'] = [rule['parents']]
-
-    options = rule.setdefault('options', {})
-
-    if isinstance(options, string_types):
-        options = options.split()
-
-    if isinstance(options, list):
-        options = dict(
-            (o[2:] if o.startswith('NO') else o, not o.startswith('NO'))
-            for o in options
-        )
-
-    rule['options'] = RoleOptions(**options)
-    return rule
-
-
-def grantrule(value):
-    if not isinstance(value, dict):
-        raise ValueError('Grant rule must be a dict.')
-    if 'acl' not in value:
-        raise ValueError('Missing acl to grant rule.')
-
-    allowed_keys = set([
-        'acl', 'database', 'schema',
-        'role', 'roles', 'role_match', 'role_attribute',
-    ])
-    defined_keys = set(value.keys())
-
-    if defined_keys - allowed_keys:
-        msg = 'Unknown parameter to grant rules: %s' % (
-            ', '.join(defined_keys - allowed_keys)
-        )
-        raise ValueError(msg)
-
-    if 'role' in value:
-        value['roles'] = value.pop('role')
-    if 'roles' in value and isinstance(value['roles'], string_types):
-        value['roles'] = [value['roles']]
-
-    if 'roles' not in value and 'role_attribute' not in value:
-        raise ValueError('Missing role in grant rule.')
-
-    return value
-
-
-def ismapping(value):
-    # Check whether a YAML value is supposed to be a single mapping.
-    if not isinstance(value, dict):
-        return False
-    return bool(set(['grant', 'ldap', 'role', 'roles']) >= set(value.keys()))
-
-
-def mapping(value):
-    # A single mapping from a query to a set of role rules. This function
-    # translate random YAML to cannonical schema.
-
-    if not isinstance(value, dict):
-        raise ValueError("Mapping should be a dict.")
-
-    if 'ldap' in value:
-        value['ldap'] = ldapquery(value['ldap'])
-
-    if 'role' in value:
-        value['roles'] = value.pop('role')
-    if 'roles' not in value:
-        value['roles'] = []
-    if not isinstance(value['roles'], list):
-        value['roles'] = [value['roles']]
-
-    value['roles'] = [rolerule(r) for r in value['roles']]
-
-    if 'grant' in value:
-        if isinstance(value['grant'], dict):
-            value['grant'] = [value['grant']]
-        value['grant'] = [grantrule(g) for g in value['grant']]
-
-    if not value['roles'] and 'grant' not in value:
-        # Don't accept unused LDAP queries.
-        raise ValueError("Missing role or grant rule.")
-
-    return value
-
-
-def syncmap(value):
-    # Validate and translate raw YAML value to cannonical form used internally.
-    #
-    # A sync map has the following canonical schema:
-    #
-    # <__all__|dbname>:
-    #   <__all__|__any__|schema>:
-    #   - ldap: <ldapquery>
-    #     roles:
-    #     - <rolerule>
-    #     - ...
-    #   ...
-    # ...
-    #
-    # But we accept a wide variety of shorthand schemas:
-    #
-    # Single mapping:
-    #
-    # roles: [<rolerule>]
-    #
-    # List of mapping:
-    #
-    # - roles: [<rolerule>]
-    # - ...
-    #
-    # dict of dbname->single mapping
-    #
-    # appdb:
-    #   roles: <rolerule>
-    #
-    # dict of dbname-> list of mapping
-    #
-    # appdb:
-    # - roles: <rolerule>
-    #
-    # dict of dbname->schema->single mapping
-    #
-    # appdb:
-    # - roles: <rolerule>
-    # dict of dbname->schema->single mapping
-    #
-    # appdb:
-    #   appschema:
-    #     roles: <rolerule>
-
-    if not value:
-        return {}
-
-    if ismapping(value):
-        value = [value]
-
-    if isinstance(value, list):
-        value = dict(__all__=value)
-
-    if not isinstance(value, dict):
-        raise ValueError("Illegal value for sync_map.")
-
-    for dbname, ivalue in value.items():
-        if ismapping(ivalue):
-            value[dbname] = ivalue = [ivalue]
-
-        if isinstance(ivalue, list):
-            value[dbname] = ivalue = dict(__any__=ivalue)
-
-        for schema, maplist in ivalue.items():
-            if isinstance(maplist, dict):
-                ivalue[schema] = maplist = [maplist]
-
-            maplist[:] = [mapping(m) for m in maplist]
-
-    return value
 
 
 class VersionAction(_VersionAction):
@@ -361,7 +153,7 @@ class Mapping(object):
 
     _auto_env = object()
 
-    def __init__(self, path, env=_auto_env, secret=False, processor=raw):
+    def __init__(self, path, env=_auto_env, secret=False, processor=V.raw):
         self.path = path
         self.arg = path.replace(':', '_')
 
@@ -475,12 +267,6 @@ class Configuration(dict):
             'binddn': None,
             'user': None,
             'password': '',
-            'default_query': {
-                'base': '',
-                'filter': '(objectClass=*)',
-                'scope': 'sub',
-                'attributes': ['cn'],
-            },
         },
         'postgres': {
             'dsn': '',
@@ -523,9 +309,9 @@ class Configuration(dict):
         ),
         Mapping('postgres:blacklist', env=None),
         Mapping('postgres:roles_query', env=None),
-        Mapping('acl_dict', processor=acldict),
+        Mapping('acl_dict', processor=V.acldict),
         Mapping('acl_groups', env=None),
-        Mapping('sync_map', env=None, processor=syncmap)
+        Mapping('sync_map', env=None, processor=V.syncmap)
     ]
 
     def __init__(self):
